+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix issue with analyzing packages nested inside git repositories on windows.
   https://github.com/dart-lang/pana/issues/1207 
+- Check the reference of the current version in `CHANGELOG.md`.
 
 ## 0.22.21